refactor(image): Migrate from asset to images
